% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert wind speeds from one unit to another}
\usage{
convert_units(wspds, inunits = "m/s", outunits = "m/s")
}
\arguments{
\item{wspds}{vector containing the wind speed observations}

\item{inunits}{units for the wind observations, "m/s" (default), "mph", "knots"}

\item{outunits}{desired units for the potential winds, "m/s" (default), "mph", "knots"}
}
\description{
Given a vector of wind speed data, this script applies a multiplier to convert from one type of measurement unit to another. Missing values are allowed.
}
\examples{
# convert from mph to m/s
convert_units(wspds, inunits = "mph")
}
\keyword{convert}
\keyword{units}
\keyword{wind}
