% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertunits.R
\name{convertunits}
\alias{convertunits}
\title{Convert wind speeds from one unit to another}
\usage{
convertunits(wspds, inunits = "m/s", outunits = "m/s")
}
\arguments{
\item{wspds}{vector containing the wind speed observations}

\item{inunits}{input units, "m/s" (default), "mph", "knots"}

\item{outunits}{desired output units, "m/s" (default), "mph", "knots"}
}
\description{
Given a vector of wind speed data, this script applies a multiplier to convert from one type of measurement unit to another. Missing values are allowed.
}
\examples{
# convert from mph to m/s
convertunits(wspds, inunits = "mph")
}
\keyword{convert}
\keyword{units}
\keyword{wind}
