% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbae.R
\name{getbae}
\alias{getbae}
\title{Calculate the bivariate absolute error for a forecast/observation wind vector pair}
\usage{
getbae(ob.u, ob.v, fcst.u, fcst.v)
}
\arguments{
\item{ob.u}{observed u component of wind}

\item{ob.v}{observed v component of wind}

\item{fcst.u}{forecast u component of wind}

\item{fcst.v}{forecast v component of wind}
}
\value{
numeric ... the bivariate absolute error
}
\description{
Calculate the bivariate absolute error for a forecast/observation wind vector pair. Missing values are not allowed and NA will be returned.
}
\examples{
getbae(ob.u, ob.v, fcst.u, fcst.v)
}
\author{
Bryan Holman
}
\keyword{absolute}
\keyword{bivariate}
\keyword{error}
\keyword{verification}
\keyword{wind}
