% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbmae.R
\name{getbmae}
\alias{getbmae}
\title{Calculate bivariate mean absolute error over many forecast/observation wind vector pairs}
\usage{
getbmae(obs.u, obs.v, fcsts.u, fcsts.v)
}
\arguments{
\item{obs.u}{vector containing observed u wind components}

\item{obs.v}{vector containing observed v wind components}

\item{fcsts.u}{vector containing forecast u wind components}

\item{fcsts.v}{vector containing forecast v wind components}
}
\description{
Calculate bivariate mean absolute error over many forecast/observation wind vector pairs.
}
\examples{
getbmae(obs.u, obs.v, fcsts.u, fcsts.v)
}
\keyword{absolute}
\keyword{bivariate}
\keyword{error}
\keyword{verification}
\keyword{wind}
