% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getemosmvrank.R
\name{getemosmvrank}
\alias{getemosmvrank}
\title{Calculate the multivariate rank given vectors of u and v ensemble forecasts}
\usage{
getemosmvrank(u, v, mu.u, mu.v, s.u, s.v, rho = 0, n = 21)
}
\arguments{
\item{u}{u-component of verifying wind observation}

\item{v}{v-component of verifying wind observation}

\item{mu.u}{u-component of EMOS mean wind vector}

\item{mu.v}{v-component of EMOS mean wind vector}

\item{s.u}{square root of u-component of EMOS variance}

\item{s.v}{square root of v-component of EMOS variance}

\item{rho}{u/v correlation coefficient}

\item{n}{number of ensemble members}
}
\value{
numeric ... the raw ensemble multivariate rank
}
\description{
Calculate the multivariate rank given vectors of u and v ensemble forecasts
}
\examples{
# calculate multivariate rank
emosmvrank <- getemosmvrank(u, v, mu.u, mu.v, s.u, s.v)
}
\author{
Bryan Holman
}
\keyword{multivariate}
\keyword{rank}
\keyword{verification}
\keyword{wind}
