% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getenorm.R
\name{getenorm}
\alias{getenorm}
\title{Calculate the two dimensional Eucledian norm}
\usage{
getenorm(x1, x2, y1, y2)
}
\arguments{
\item{x1}{point 1 x component}

\item{x2}{point 2 x component}

\item{y1}{point 1 y component}

\item{y2}{point 2 y component}
}
\description{
Calculates the two dimensional Eucledian norm between two points.
}
\examples{
wspdwdir <- getwspdwdir(wspd, wdir)
}
\keyword{eucledian}
\keyword{normal}
\keyword{verification}
\keyword{wind}
