% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getes.R
\name{getes}
\alias{getes}
\title{Calculate the energy score given vectors of u and v ensemble forecasts}
\usage{
getes(ob.u, ob.v, ens.us, ens.vs)
}
\arguments{
\item{ob.u}{u-component of verifying wind observation}

\item{ob.v}{v-component of verifying wind observation}

\item{ens.us}{u-components of ensemble wind forecasts}

\item{ens.vs}{v-components of ensemble wind forecasts}
}
\value{
numeric ... the raw ensemble energy score
}
\description{
Calculate the energy score given vectors of u and v ensemble forecasts
}
\examples{
# the verifying observation
u <- 4.5
v <- -3.6

# synthetic ensembles
ens.us <- c(4.6, 3.5, 2.9, 5.3)
ens.vs <- c(-4.8, -2.1, -3.9, -3.0)

# calculate energy score
es <- getes(u, v, ens.us, ens.vs)
}
\author{
Bryan Holman
}
\keyword{energy}
\keyword{score}
\keyword{verification}
\keyword{wind}
