% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getesapprox.R
\name{getesapprox}
\alias{getesapprox}
\title{Calculate the approximate energy score given EMOS parameters}
\usage{
getesapprox(u, v, mu.u, mu.v, s.u, s.v, rho = 0, n = 10000)
}
\arguments{
\item{u}{u-component of verifying wind observation}

\item{v}{v-component of verifying wind observation}

\item{mu.u}{u-component of EMOS mean wind vector}

\item{mu.v}{v-component of EMOS mean wind vector}

\item{s.u}{square root of u-component of EMOS variance}

\item{rho}{u/v correlation coefficient}

\item{n}{number of distribution samples used to approximate energy score}

\item{s.u}{square root of v-component of EMOS variance}
}
\value{
numeric ... the approximate energy score
}
\description{
Calculate the approximate energy score given EMOS parameters
}
\examples{
NULL
}
\author{
Bryan Holman
}
\keyword{energy}
\keyword{score}
\keyword{verification}
\keyword{wind}
