% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getuv.R
\name{getuv}
\alias{getuv}
\title{Convert wind speed and wind direction into u and v wind components}
\usage{
getuv(wspd, wdir)
}
\arguments{
\item{wspd}{wind speed value}

\item{wdir}{wind direction value}
}
\description{
Convert wind speed and wind direction into u and v wind components
}
\examples{
wspd <- 6.0
wdir <- 147
uv <- getuv(wspd, wdir)
}
\keyword{convert}
\keyword{direction}
\keyword{speed}
\keyword{u}
\keyword{v}
\keyword{wind}
