% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwdirerror.R
\name{getwdirerror}
\alias{getwdirerror}
\title{Calculate the wind direction error of a forecast/observation pair}
\usage{
getwdirerror(ob.wdir, fcst.wdir)
}
\arguments{
\item{ob.wdir}{observed wind direction}

\item{fcst.wdir}{forecasted wind direction}
}
\description{
Calculate the wind direction error of a forecast/observation pair
}
\examples{
ob.wdir <- 63
fcst.wdir <- 147
wdirerror <- getwdirerror(wspd, wdir)
}
\keyword{direction}
\keyword{error}
\keyword{wind}
