% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwspdwdir.R
\name{getwspdwdir}
\alias{getwspdwdir}
\title{Convert u and v wind components into wind speed and wind direction}
\usage{
getwspdwdir(u, v)
}
\arguments{
\item{u}{zonal wind component}

\item{v}{meridional wind component}
}
\description{
Convert u and v wind components into wind speed and wind direction
}
\examples{
u <- -4.56
v <- 147
wspdwdir <- getwspdwdir(wspd, wdir)
}
\keyword{convert}
\keyword{direction}
\keyword{speed}
\keyword{u}
\keyword{v}
\keyword{wind}
