% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndbc_10m.R
\name{ndbc_10m}
\alias{ndbc_10m}
\title{Calculate the 10-m winds for a vector of wind speed measurements according to Hsu et al. (1994)}
\usage{
ndbc_10m(wspds, zm, zref = 10, inunits = "m/s", outunits = "m/s",
  to.na = TRUE, missing = -99999)
}
\arguments{
\item{wspds}{vector containing the wind speed observations}

\item{zm}{the measurement height for wind speed observations (in m)}

\item{zref}{the reference height (the height we want to estimate the winds), defaults to 10-m}

\item{inunits}{units for the wind observations, "m/s" (default), "mph", "knots"}

\item{outunits}{desired units for the potential winds, "m/s" (default), "mph", "knots"}

\item{to.na}{turn large values to NA (default = TRUE)}

\item{missing}{specify missing value (if to.na = FALSE)}
}
\description{
This function will return the 10-m winds given vectors of wind observations and the measurement height.
}
\examples{
# convert from mph to m/s
ndbc_10m(wspds, 10, inunits = "mph")

# get potential winds from a station with anemometer height of 4.9 m
ndbc_10m(wspds, 4.9)
}
\keyword{potential}
\keyword{wind}
