% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winds_10m.R
\name{winds_10m}
\alias{winds_10m}
\title{Calculate the 10-m winds for a vector of wind speed measurements}
\usage{
winds_10m(bins, z0, wspds, wdirs, zm, zref = 10, inunits = "m/s",
  outunits = "m/s", to.na = TRUE, missing = -99999)
}
\arguments{
\item{bins}{vector containing the cutoffs of the wind direction bins}

\item{z0}{vector containing the roughness lengths associated with each wind direction bin}

\item{wspds}{vector containing the wind speed observations}

\item{wdirs}{vector containing the wind direction observations}

\item{zm}{the measurement height for wind speed observations (in m)}

\item{zref}{the reference height (the height we want to estimate the winds), defaults to 10-m}

\item{inunits}{units for the wind observations, "m/s" (default), "mph", "knots"}

\item{outunits}{desired units for the potential winds, "m/s" (default), "mph", "knots"}

\item{to.na}{turn large values to NA (default = TRUE)}

\item{missing}{specify missing value (if to.na = FALSE)}
}
\description{
This function will return the 10-m winds given vectors of wind observations (wind speed and direction), a vector containing the high end of the wind direction bins, the roughness lengths associated with each wind direction bins, and the measurement height of the wind observations.
}
\examples{
# convert from mph to m/s
winds_10m(bins, z0, wspd, wdir, zm, inunits = "mph")
}
\keyword{potential}
\keyword{wind}
